requirements.txt
torch
torchvision
Pillow
scikit-learn
matplotlib
streamlit
pandas

